docs: refactor zenoh-bridge architecture and fix mkdocs configuration (#24)
* feat: zenoh-bridge migrating containers to cloud-edge model

* docs: Add zenoh-bridge deployment document

* docs: minor: fix wrong item and link

* docs: Update zenoh-bridge documents

complying with current directory structure

Signed-off-by: Tzu-Ching Yang <shingekinocore@gmail.com>

* docs: comply with mkdocs awesome-page style

Signed-off-by: Tzu-Ching Yang <shingekinocore@gmail.com>

* docs: fix mkdocs `.page` issue

This commit is to fix all the issue w.r.t. mkdocs.

`.page` is excluded so it will not be applied in the top tabs.

The old solution to maintain tab order is to add nav directly
in mkdocs.yaml, which will disble awesome-pages. same-dir also
is not compatible with awesome-pages.

Meanwhile, the non-existing pages mention in
quick-start/start-quide/.pages cause awesome-pages not to work as
expected. So I comment them out.

Signed-off-by: Tzu-Ching Yang <shingekinocore@gmail.com>

* docs: adjust zenoh-bridge doc

No redundant file remains.

Signed-off-by: Tzu-Ching Yang <shingekinocore@gmail.com>

* docs: update zenoh-bridge README.md link

Signed-off-by: Tzu-Ching Yang <shingekinocore@gmail.com>

* Merge remote-tracking branch 'origin/main' into feat/zenoh-bridge/cloud-edge

Signed-off-by: Tzu-Ching Yang <shingekinocore@gmail.com>

---------

Signed-off-by: Tzu-Ching Yang <shingekinocore@gmail.com>